[exporter/stefexporter] Add basic STEF exporter implementation
#### Description

Added STEF exporter implementation for metrics, sending data over gRPC stream.
For now only queuing and retry exporter helpers are used. We will need to
decide later if other helpers are needed for this exporter.

#### Testing

Unit tests that verify connecting, reconnecting, sending, acking of data
are included.

#### Documentation

Added to README.

#### Future Work

More extensive test coverage is desirable and will likely be added
in the future.

We likely want to implement STEF receiver and add STEF as a tested protocol
to our testbed.